labs/1-2-emulator/project/backend/requirements.txt
pytest
fastapi
loguru
uvicorn[standard]
opencv-python-headless
python-multipart
numpy
sqlalchemy
psycopg2-binary
PyJWT
python-dotenv
celery
redis
